+++ .pylintrc
@@ -74,30 +74,22 @@ disable=
   broad-except,
   cell-var-from-loop,
   consider-iterating-dictionary,
-  cyclic-import,
   dangerous-default-value,
-  duplicate-code,
-  exec-used,
   fixme,
   global-statement,
-  inherit-non-class,
   invalid-name,
   len-as-condition,
   locally-disabled,
   logging-not-lazy,
   missing-docstring,
-  multiple-imports,
   no-else-return,
   no-init,
   no-member,
   no-self-use,
-  not-context-manager,
   protected-access,
   redefined-builtin,
   redefined-outer-name,
   redefined-variable-type,
-  relative-import,
-  signature-differs,
   super-init-not-called,
   too-few-public-methods,
   too-many-arguments,
@@ -116,14 +108,13 @@ disable=
   unused-argument,
   unused-variable,
   wildcard-import,
-  wrong-import-order,
   wrong-import-position,
   deprecated-lambda,
   simplifiable-if-statement,
   unidiomatic-typecheck,
   global-at-module-level,
-  global-variable-not-assigned
-
+  inconsistent-return-statements,
+  keyword-arg-before-vararg
 
 [REPORTS]
